Moved ProtoBufRegistry and ProtoBufSerializer to the geotrellis sub-package
