+++ geopyspark/geotrellis/catalog.py
@@ -5,7 +5,7 @@ import json
 from collections import namedtuple
 from urllib.parse import urlparse
 
-from geopyspark.protobufregistry import multibandtile_decoder
+from geopyspark.geotrellis.protobufcodecs import multibandtile_decoder
 from geopyspark.geotrellis import Metadata, Extent
 from geopyspark.geotrellis.rdd import TiledRasterRDD
 from geopyspark.geotrellis.constants import TILE, ZORDER, SPATIAL
