+++ geopyspark/tests/schema_tests/multiband_schema_test.py
@@ -5,8 +5,8 @@ import pytest
 from pyspark import RDD
 from pyspark.serializers import AutoBatchedSerializer
 from geopyspark.protobuf import tileMessages_pb2
-from geopyspark.protobufserializer import ProtoBufSerializer
-from geopyspark.protobufregistry import multibandtile_decoder, multibandtile_encoder
+from geopyspark.geotrellis.protobufserializer import ProtoBufSerializer
+from geopyspark.geotrellis.protobufcodecs import multibandtile_decoder, multibandtile_encoder
 from geopyspark.tests.base_test_class import BaseTestClass
