+++ geopyspark/tests/schema_tests/projected_extent_schema_test.py
@@ -4,9 +4,10 @@ import pytest
 from pyspark import RDD
 from pyspark.serializers import AutoBatchedSerializer
 from geopyspark.geotrellis import Extent, ProjectedExtent
-from geopyspark.protobufserializer import ProtoBufSerializer
-from geopyspark.protobufregistry import (projected_extent_decoder, projected_extent_encoder,
-                                         to_pb_projected_extent)
+from geopyspark.geotrellis.protobufserializer import ProtoBufSerializer
+from geopyspark.geotrellis.protobufcodecs import (projected_extent_decoder,
+                                                  projected_extent_encoder,
+                                                  to_pb_projected_extent)
 from geopyspark.tests.base_test_class import BaseTestClass
